backend_root/requirements.txt
beautifulsoup4==4.11.1
Django==4.1.5
djangorestframework==3.14.0
dash
plotly
